+++ database/migrations/2020_06_26_081202_create_thfundmasters_table.php
@@ -16,14 +16,14 @@ class CreateThfundmastersTable extends Migration
         Schema::create('thfundmasters', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->integer('member_id');
-            $table->decimal('op_balance')->nullable();
+            $table->float('op_balance')->nullable();
             $table->date('op_balance_date')->nullable();
 
             $table->integer('total_month')->nullable();
-            $table->decimal('total_collection')->nullable();
-            $table->decimal('total_interest')->nullable();
+            $table->float('total_collection')->nullable();
+            $table->float('total_interest')->nullable();
             
-            $table->decimal('cl_balance')->nullable();
+            $table->float('cl_balance')->nullable();
             $table->date('cl_balance_date')->nullable();   
             
             $table->string('status')->nullable;
