+++ database/migrations/2020_06_26_081416_create_thfundspecifications_table.php
@@ -16,7 +16,7 @@ class CreateThfundspecificationsTable extends Migration
         Schema::create('thfundspecifications', function (Blueprint $table) {
             $table->bigIncrements('id');
 
-            $table->decimal('monthly_premium');
+            $table->float('monthly_premium');
             $table->integer('yearly_roi');         
 
             $table->date('date_of_activation');
